ninja_api_server/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
contextlib2==21.6.0
cryptography==41.0.1
Django==4.2.2
django-environ==0.10.0
django-ninja==0.22.2
django-ninja-extra==0.19.1
django-ninja-jwt==5.2.5
h11==0.14.0
idna==3.4
injector==0.20.1
pycparser==2.21
pydantic==1.10.9
PyJWT==2.7.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
